refactor: XYNE-16525 move enums to types.ts and sync to Prisma parity
Groundwork for the enum -> TEXT migration. Two changes, no behaviour change:

1. Extract every enum / literal-union definition out of zero/schema.ts into a
   new zero/types.ts, leaving schema.ts as purely the Zero table schema (its
   generated purpose). schema.ts re-exports types.ts, so every existing import
   path (`@xyne/shared`, `../zero/schema`) keeps resolving unchanged.

2. Bring the shared enums to exact parity with prisma/schema.prisma, since they
   become the single source of truth for allowed values once the DB columns are
   plain TEXT and the enum constraint is gone.

   Added 16 enums that only existed in Prisma: AppPermissionStatus,
   AppPermissionType, NudgeType, QueryType, RecordingStatus,
   SessionRecordingProcessStatus, TagMethod, TeamIntelligenceBatchStatus,
   TeamIntelligenceUserIngestionStatus, TelepresenceDeviceType,
   TelepresenceHealthStatus, VespaInsertionStatus, VespaOperationType,
   WorkflowEventType, WorkflowExecutionMode, WorkflowMappingEntityType.

   Added values that had drifted out of sync: AttachmentEntityType.WORKFLOW_STEPS,
   ExternalEntityType.{ATTACHMENT,CANVAS}, NotificationDeliveryMethod.{IOS,ANDROID},
   NotificationStatus.{DELIVERED,FAILED}, PRStatus.UPDATED, and 12 NotificationType
   values (call/read/email-fetch/canvas/recording).

All 110 enums used as Prisma model field types now match their Prisma
definition value-for-value.
